crawler-temu-seller-kuajingmaihuo-data/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
crypto==1.4.1
cryptography==43.0.3
DBUtils==3.1.0
ddddocr==1.5.6
DigiCore==1.1.9
dnspython==2.6.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.12.23
frozenlist==1.4.1
fsspec==2024.12.0
h11==0.14.0
humanfriendly==10.0
idna==3.7
imageio==2.36.1
Jinja2==3.1.5
lazy_loader==0.4
loguru==0.7.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.0.5
Naked==0.1.32
networkx==3.2.1
numpy==2.0.1
onnxruntime==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.6
packaging==24.2
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==11.0.0
protobuf==5.29.3
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyExecJS==1.5.1
pymongo==4.8.0
PyMySQL==1.1.1
pypdfium2==4.30.0
pysmb==1.2.9.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
tifffile==2024.8.30
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.0
win32-setctime==1.1.0
yarl==1.9.4
